Move app logic to library crate
